+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Develop
 
+- Send `CLICK` event if there is an overlap between max time between clicks and new click arrives
+
 ## v0.0.2
 
 - Add `LWBTN_CFG_GET_STATE_MODE` to control *get state* mode